+++ src/main/java/me/Fupery/ArtMap/Compatability/DenizenCompat.java
@@ -3,11 +3,11 @@ package me.Fupery.ArtMap.Compatability;
 import java.lang.reflect.Field;
 import java.util.Arrays;
 
-import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
 import org.bukkit.plugin.Plugin;
 
 import io.netty.channel.Channel;
+import me.Fupery.ArtMap.ArtMap;
 import me.Fupery.ArtMap.Compatability.Dipenizen.ArtMapArt;
 import me.Fupery.ArtMap.Compatability.Dipenizen.ArtMapArtist;
 import me.Fupery.ArtMap.Compatability.Dipenizen.ArtMapArtists;
@@ -19,13 +19,13 @@ public class DenizenCompat implements ReflectionHandler {
     private boolean loaded = false;
 
     DenizenCompat() {
-        Plugin plugin = Bukkit.getPluginManager().getPlugin("Denizen");
+        Plugin plugin = ArtMap.instance().getServer().getPluginManager().getPlugin("Denizen");
         loaded = (plugin != null && plugin.isEnabled());
         if(!loaded){
             return;
         }
         // Add denizen objects for use in scripts
-        Plugin dipenizen = Bukkit.getPluginManager().getPlugin("Dipenizen");
+        Plugin dipenizen = ArtMap.instance().getServer().getPluginManager().getPlugin("Dipenizen");
         if(dipenizen != null && dipenizen.isEnabled()) {
             ObjectFetcher.registerWithObjectFetcher(ArtMapArt.class);
             ObjectFetcher.registerWithObjectFetcher(ArtMapArtist.class);